backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Data processing and visualization
requests==2.31.0
pandas==2.1.4
matplotlib==3.8.2
mplfinance==0.12.9b7

# Cache and file handling
redis==5.0.1
aiofiles==23.2.1